fraud_detector/requirements.txt
numpy==1.26.4
pandas==2.2.3
kafka-python==2.0.2
catboost==1.2.5
